fixes a bug where binary expressions were not being processed properly with numbers due to a type difference between the interpreter check and the token value

+++ src/com/craftinginterpreters/lox/Scanner.java
@@ -109,7 +109,7 @@ public class Scanner {
             while(isDigit(peek())) advance();
             addToken(NUMBER, Double.parseDouble((source.substring(start, current))));
         } else {
-            addToken(NUMBER, Integer.parseInt((source.substring(start, current))));
+            addToken(NUMBER, Double.parseDouble((source.substring(start, current))));
         }
     }
